# Conflicts:
#	Client/Client.cs
#	Core/Maps/Geometry/SectorFlat.cs
#	Core/Render/OpenGL/GLRenderer.cs
#	Core/Render/OpenGL/Renderers/World/WorldRenderableGeometry.cs
#	Core/Render/OpenGL/Renderers/World/WorldRenderer.cs
#	Core/Render/OpenGL/Texture/GLTextureManager.cs
#	Core/Render/OpenGL/Util/GLCapabilities.cs
#	Core/Resources/ResourceTracker.cs
#	Core/Util/Console.cs
#	Core/Util/Constants.cs

+++ Core/Render/OpenGL/Texture/TextureInfo.cs
@@ -11,8 +11,8 @@ namespace Helion.Render.OpenGL.Texture
         public readonly float TopV;
         public readonly float InverseU;
         public readonly float InverseV;
-        public readonly int Width;
-        public readonly int Height;
+        public readonly float Width;
+        public readonly float Height;
 
         public TextureInfo(GLTexture texture)
         {
